# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TestPilot AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organizations/current/billing": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Current Billing",
        "operationId": "get_current_billing_api_v1_organizations_current_billing_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/issues": {
      "get": {
        "tags": [
          "issues"
        ],
        "summary": "List Issues",
        "operationId": "list_issues_api_v1_projects__project_id__issues_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssuePriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "issues"
        ],
        "summary": "Create Issue",
        "operationId": "create_issue_api_v1_projects__project_id__issues_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/issues/from-result/{result_id}": {
      "post": {
        "tags": [
          "issues"
        ],
        "summary": "Create Issue From Result",
        "operationId": "create_issue_from_result_api_v1_projects__project_id__issues_from_result__result_id__post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueFromResultRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/issues/{issue_id}": {
      "get": {
        "tags": [
          "issues"
        ],
        "summary": "Get Issue",
        "operationId": "get_issue_api_v1_projects__project_id__issues__issue_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Issue Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "issues"
        ],
        "summary": "Update Issue",
        "operationId": "update_issue_api_v1_projects__project_id__issues__issue_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Issue Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/issues/{issue_id}/attachments": {
      "post": {
        "tags": [
          "issues"
        ],
        "summary": "Add Attachment",
        "description": "Accepts either a multipart file upload or a JSON `{storage_key, type}`\nbody referencing an existing artifact (contracts/issues-api.md) \u2014 the\ntwo variants share one endpoint, dispatched on the request's own\nContent-Type, since FastAPI/OpenAPI has no single-route way to declare\nboth a file body and a JSON body as alternatives.",
        "operationId": "add_attachment_api_v1_projects__project_id__issues__issue_id__attachments_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Issue Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueAttachmentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingDetail": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/PlanSummary"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageSummary"
          },
          "period": {
            "$ref": "#/components/schemas/BillingPeriod"
          }
        },
        "type": "object",
        "required": [
          "plan",
          "usage",
          "period"
        ],
        "title": "BillingDetail"
      },
      "BillingPeriod": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "BillingPeriod"
      },
...
      "CreateIssueFromResultRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "priority"
        ],
        "title": "CreateIssueFromResultRequest",
        "description": "`title`/`description` default from the failure when omitted (FR-087)."
      },
      "CreateIssueRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "severity",
          "priority"
        ],
        "title": "CreateIssueRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IssueAttachmentPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/IssueAttachmentType"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "url",
          "created_at"
        ],
        "title": "IssueAttachmentPublic"
      },
      "IssueAttachmentType": {
        "type": "string",
        "enum": [
          "screenshot",
          "log"
        ],
        "title": "IssueAttachmentType"
      },
      "IssueDetailResponse": {
        "properties": {
          "issue": {
            "$ref": "#/components/schemas/IssuePublic"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/IssueAttachmentPublic"
            },
            "type": "array",
            "title": "Attachments"
          },
          "source_test_case": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestCaseRefPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_test_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestRunRefPublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "issue",
          "attachments",
          "source_test_case",
          "source_test_run"
        ],
        "title": "IssueDetailResponse"
      },
      "IssueListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssuePublic"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "IssueListResponse"
      },
      "IssuePriority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "IssuePriority",
        "description": "Same value set as testcases.models.TestCasePriority, kept separate\nfor the same reason as IssueSeverity above."
      },
      "IssuePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "assignee_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee User Id"
          },
          "source_test_case_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Test Case Id"
          },
          "source_test_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Test Run Id"
          },
          "created_by_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Created By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "description",
          "severity",
          "priority",
          "status",
          "assignee_user_id",
          "source_test_case_id",
          "source_test_run_id",
          "created_by_user_id",
          "created_at",
          "updated_at"
        ],
        "title": "IssuePublic"
      },
      "IssueSeverity": {
        "type": "string",
        "enum": [
          "minor",
          "major",
          "critical",
          "blocker"
        ],
        "title": "IssueSeverity",
        "description": "Same value set as testcases.models.TestCaseSeverity \u2014 kept separate\nsince an issue's severity is independently editable after creation, not\nderived from its source test case (same reasoning ai_analysis.models.\nAnalysisSeverity documents for its own, otherwise-identical, enum)."
      },
      "IssueStatus": {
        "type": "string",
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "closed",
          "wont_fix"
        ],
        "title": "IssueStatus",
        "description": "Declared open-to-wont_fix in the lifecycle order contracts/issues-api.md\ndescribes (open -> in_progress -> resolved -> closed, or open/in_progress\n-> wont_fix) \u2014 see issues/service.py's transition-validation for the\nactual state-machine rules this ordering alone doesn't encode."
      },
...
      "PlanSummary": {
        "properties": {
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "limits": {
            "$ref": "#/components/schemas/PlanLimits"
          }
        },
        "type": "object",
        "required": [
          "tier",
          "limits"
        ],
        "title": "PlanSummary"
      },
...
      "UpdateIssueRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueSeverity"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssuePriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignee_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee User Id"
          }
        },
        "type": "object",
        "title": "UpdateIssueRequest"
      },
...
      "UsageSummary": {
        "properties": {
          "projects": {
            "type": "integer",
            "title": "Projects"
          },
          "test_executions": {
            "type": "integer",
            "title": "Test Executions"
          },
          "ai_operations": {
            "type": "integer",
            "title": "Ai Operations"
          },
          "members": {
            "type": "integer",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "projects",
          "test_executions",
          "ai_operations",
          "members"
        ],
        "title": "UsageSummary"
      },
...
    }
  }
}